fix(accountsdb): use current slot during rollback if off by one
In case if the requested slot, to rollback to, is older than the current
slot just by one, then return the current slot. This solves the bug when
we try to rollback to older slot when no snapshots exists and validator
fails to start. Additionally this is an expected scnerio during
validator restarts, since ledger is always behind accountsdb by at least
one slot.

closes gh-336