feat(config): derive merge (#476)
Closes #471 

Create a `Mergeable` derive trait that automatically merges the fields
with default values with the non-default fields of another instance. It
also handles nested Mergeable structs.

Note: in order to handle nested Mergeable structs, it uses the heuristic
that nested mergeable struct have a type name containing `Config`. This
is not a very robust method, but it works as we only use it for internal
config structs.