fix: refactor Stop the World locking behavior
since accountsdb snapshots take place in async context
(due to the slot advancement happening in tickers),
we end up blocking the runtime, while trying to acquire
a sync lock. In order to avoid that, this PR moves the
whole lock acquisition and snapshot logic a separate thread
and performs it in the background, which makes slot advancement
a non-blocking operation.