feat: schedule tasks (#493)
Closes #492

Enables scheduling static periodic tasks (repeats the exact same
instruction a fixed number of times) in the ephemeral rollup. Refer to
the
[MIMD](https://github.com/magicblock-labs/magicblock-validator/discussions/479)
for detailed explanations.

# Changes

## Task scheduler

A new component that starts along the validator. It listens for changes
to a newly introduced special `TaskContext` account. A new
`TaskSchedulerConfig` is introduced to set its parameters.

## Magic program

Introduces two new Magic program instructions:
- `ScheduleTask` creates a new task that will be executed periodically.
This instruction can only be called via CPI. When scheduling the task,
all accounts used by the task must be signers of the transaction or
owned by the program that is calling it. This is to ensure that the
scheduled instructions are valid, as we do not verify the signatures of
scheduled tasks.
- `CancelTask` cancels an existing task.